bot_project/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-phonenumber-field==7.1.0
django-timezone-field==5.1
djangorestframework==3.14.0
django-import_export
gunicorn==21.2.0
idna==3.4
eventlet
kombu==5.3.1
packaging==23.1
phonenumbers==8.13.16
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
vk-api==11.9.9
wcwidth==0.2.6
djangorestframework==3.14.0
django-object-actions==4.1.0
